+++ app/src/main/java/com/kusitms/assignmentandroid/retrofit/RetrofitAPI.java
@@ -1,11 +1,10 @@
 package com.kusitms.assignmentandroid.retrofit;
 
-import com.kusitms.assignmentandroid.dto.LoginResult;
-import com.kusitms.assignmentandroid.dto.SerialNumberResult;
+import com.kusitms.assignmentandroid.retrofit.dto.LoginResult;
+import com.kusitms.assignmentandroid.retrofit.dto.SerialNumberResult;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
-import retrofit2.http.Header;
 import retrofit2.http.Query;
 
 public interface RetrofitAPI {
