+++ app/src/main/res/layout/fragment_home.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".HomeFragment">
+    tools:context=".view.HomeFragment">
 
     <TextView
         android:id="@+id/tvName"
