+++ app/src/main/res/layout/fragment_order.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".OrderFragment">
+    tools:context=".view.OrderFragment">
 
     <TextView
         android:id="@+id/textView"
